iOS examples. Returned the ContentView back to how it was
