+++ examples/codeviewer/iosApp/iosApp/ContentView.swift
@@ -4,8 +4,11 @@ import shared
 
 struct ContentView: View {
     var body: some View {
-        ComposeView()
-            .ignoresSafeArea(.keyboard) // Compose has own keyboard handler
+        ZStack {
+            Color(#colorLiteral(red: 0.235, green: 0.247, blue: 0.255, alpha: 1)).ignoresSafeArea(.all)
+            ComposeView()
+                .ignoresSafeArea(.all, edges: .bottom) // Compose has own keyboard handler
+        }.preferredColorScheme(.dark)
     }
 }
