+++ examples/interop/ios-compose-in-swiftui/iosApp/iosApp/ContentView.swift
@@ -1,0 +1,17 @@
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        TabView {
+            ComposeInSwiftUIScreen()
+                .tabItem {
+                    Label("Compose in SwiftUI", systemImage: "star.fill")
+                }
+
+            YetAnotherSwiftUIScreen()
+                .tabItem {
+                    Label("SwiftUI", systemImage: "gear")
+                }
+        }
+    }
+}
