+++ examples/interop/ios-compose-in-swiftui/iosApp/iosApp/iOSApp.swift
@@ -4,17 +4,7 @@ import SwiftUI
 struct iOSApp: App {
     var body: some Scene {
         WindowGroup {
-            TabView {
-                ComposeInSwiftUIScreen()
-                    .tabItem {
-                        Label("Compose in SwiftUI", systemImage: "star.fill")
-                    }
-
-                YetAnotherSwiftUIScreen()
-                    .tabItem {
-                        Label("SwiftUI", systemImage: "gear")
-                    }
-            }
+            ContentView()
         }
     }
 }
