+++ examples/interop/ios-swiftui-in-compose/iosApp/iosApp/ContentView.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+import shared
+import UIKit
+
+struct ContentView: View {
+    var body: some View {
+        ComposeViewControllerRepresentable()
+            .ignoresSafeArea(.all)
+    }
+}
