+++ examples/interop/ios-swiftui-in-compose/iosApp/iosApp/iOSApp.swift
@@ -1,13 +1,10 @@
 import SwiftUI
-import shared
-import UIKit
 
 @main
 struct iOSApp: App {
     var body: some Scene {
         WindowGroup {
-            ComposeViewControllerRepresentable()
-                .ignoresSafeArea(.all)
+            ContentView()
         }
     }
 }
