+++ examples/interop/ios-uikit-in-compose/iosApp/iosApp/iOSApp.swift
@@ -1,22 +1,10 @@
 import SwiftUI
-import shared
-import UIKit
 
 @main
 struct iOSApp: App {
     var body: some Scene {
         WindowGroup {
-            ComposeViewControllerToSwiftUI()
-                .ignoresSafeArea(.all)
+            ContentView()
         }
-    }
-}
-
-struct ComposeViewControllerToSwiftUI: UIViewControllerRepresentable {
-    func makeUIViewController(context: Context) -> UIViewController {
-        return Main_iosKt.ComposeEntryPoint()
-    }
-
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
     }
 }
