+++ examples/todoapp-lite/iosApp/iosApp/ContentView.swift
@@ -2,6 +2,7 @@ import UIKit
 import SwiftUI
 import shared
 
+
 struct ContentView: View {
     var body: some View {
         ComposeView()
@@ -14,5 +15,6 @@ struct ComposeView: UIViewControllerRepresentable {
         Main_iosKt.MainViewController()
     }
 
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
+    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
+    }
 }
