+++ examples/widgets-gallery/iosApp/iosApp/iOSApp.swift
@@ -1,22 +1,10 @@
-import UIKit
 import SwiftUI
-import shared
 
 @main
 struct iOSApp: App {
     var body: some Scene {
         WindowGroup {
-            ComposeView()
-                .ignoresSafeArea(.keyboard) // Compose has own keyboard handler
+            ContentView()
         }
-    }
-}
-
-struct ComposeView: UIViewControllerRepresentable {
-    func makeUIViewController(context: Context) -> UIViewController {
-        Main_iosKt.MainViewController()
-    }
-
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
     }
 }
